requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
tensorflow
python-dotenv
torch
uvicorn
asyncio